Class Info Box with edit, cancel, and save capabilities.
Summary:
Created a ClassInfo component on the classroom page. Edit button allows the classroom name and the date range to be modified. EditableDateRange component allows start and end date to be changed if the ClassInfo box is editable. Cancel toggles back to an uneditable state, and the save button saves the current input and returns to an uneditable state.
{F4242}
{F4243}
{F4244}

Test Plan: Start the rails server and route to any classroom page (/classrooms/1). Click Edit and change the classroom name and the date range. Click Cancel and the page reverts back to the original data. Click Edit again and edit the fields, and click Save - the information on the page updates to reflect the new inputs.

Reviewers: nnarayen

Reviewed By: nnarayen

Subscribers: bigblue, nnarayen, liang, tofu, generic

Projects: #bizworld

Differential Revision: http://phab.calblueprint.org/D77