DeleteClassroomModal component to delete a classroom.
Summary:
Create drop down modal to delete a classroom and re-route to the classrooms page.
{F6654}
{F6655}
{F6656}

Test Plan: Route to 'teachers/1/classrooms' and create a classroom. Click Delete Classroom at the bottom of the Students Table and you will be taken back to the teacher's classrooms page, and the deleted classroom will not appear.

Reviewers: nnarayen

Reviewed By: nnarayen

Subscribers: bigblue, liang, tofu, generic

Projects: #bizworld

Differential Revision: http://phab.calblueprint.org/D102