Added ability for admins to set correct MC choice
Summary:
Before this diff, an admin would not be able to set/change the correct
answer to a multiple choice question. With this diff, admins can now
do this.

Also: super minor updates to README, gitignore, and seed scripts

Test Plan:
Try to change the correct answer to a question, should
work now.

Reviewers: nnarayen

Reviewed By: nnarayen

Subscribers: sam, bigblue, liang, tofu, generic, caseytaco

Projects: #bizworld

Differential Revision: http://phab.calblueprint.org/D122