Implemented adding/removing MC question options
Summary:
Diff involves a somewhat substantial refactor of Question components.

1. AdminQuestions are now rendered entirely by their state instead of
their props. `this.state.question` for an AdminQuestion is initialized
as a copy of `this.props.question` in the constructor and all following
state changes are kept local to the AdminQuestion until save/delete.
2. On save or delete, the AdminQuestion triggers a callback passed as a
prop by the parent form component AdminFormQuestions. This callback
needs to be passed in because the parent form component needs to
calculate the proper insertion index for new questions at the time of
save.
3. And other stuff (see commit contents)

Test Plan: arc patch

Reviewers: nnarayen, caseytaco, liang

Reviewed By: caseytaco, liang

Subscribers: bigblue, tofu, generic, caseytaco

Projects: #bizworld

Differential Revision: http://phab.calblueprint.org/D450